requirements.txt
fastapi~=0.115
uvicorn~=0.34rea
openai>=1.6
python-dotenv~=1.1
qdrant-client~=1.9
langchain~=0.2
langchain-community~=0.2